+++ src/main/java/pvpmode/ChatUtils.java
@@ -57,9 +57,12 @@ public class ChatUtils
     {
         for (String message : messages)
         {
-            ChatComponentText text = new ChatComponentText (message);
-            text.getChatStyle ().setColor (color);
-            recipient.addChatMessage (text);
+            for (String line : message.split ("\n"))
+            {
+                ChatComponentText text = new ChatComponentText (line);
+                text.getChatStyle ().setColor (color);
+                recipient.addChatMessage (text);
+            }
         }
     }
 
@@ -82,11 +85,14 @@ public class ChatUtils
     {
         for (String message : messages)
         {
-            ChatComponentText prefix = new ChatComponentText (
+            for (String line : message.split ("\n"))
+            {
+                ChatComponentText prefix = new ChatComponentText (
                 PvPMode.prefixGlobalMessages ? PvPMode.globalMessagePrefix : "");
-            ChatComponentText text = new ChatComponentText (message);
-            text.getChatStyle ().setColor (color);
-            PvPMode.cfg.sendChatMsg (prefix.appendSibling (text));
+                ChatComponentText text = new ChatComponentText (line);
+                text.getChatStyle ().setColor (color);
+                PvPMode.cfg.sendChatMsg (prefix.appendSibling (text));
+            }
         }
     }