+++ src/main/java/pvpmode/command/AbstractPvPCommand.java
@@ -1,6 +1,9 @@
 package pvpmode.command;
 
+import java.util.Collection;
+
 import org.apache.commons.lang3.ArrayUtils;
+import org.apache.commons.lang3.tuple.Triple;
 
 import net.minecraft.command.*;
 import net.minecraft.entity.player.EntityPlayer;
@@ -8,6 +11,25 @@ import pvpmode.*;
 
 public abstract class AbstractPvPCommand extends CommandBase
 {
+
+    /**
+     * Returns a short description for this command (or every sub-command). The
+     * triple contains firstly the (sub-)command name, then the command usage and
+     * finally the short help message.
+     */
+    public abstract Collection<Triple<String, String, String>> getShortHelpMessages (ICommandSender sender);
+
+    /**
+     * Returns a detailed description for this command (or every sub-command). The
+     * triple contains firstly the (sub-)command name, then the command usage and
+     * finally the detailed help message.
+     */
+    public abstract Collection<Triple<String, String, String>> getLongHelpMessages (ICommandSender sender);
+
+    /**
+     * Returns a help message regarding the command in general.
+     */
+    public abstract String getGeneralHelpMessage (ICommandSender sender);
 
     protected void usageError (ICommandSender sender)
     {
@@ -61,7 +83,7 @@ public abstract class AbstractPvPCommand extends CommandBase
     protected void requireNonPvPSender (EntityPlayer sender)
     {
         if (PvPUtils.isInPvP (sender))
-            throw new CommandException ("You cannot use this command while in PvP");
+            throw new CommandException ("You cannot use this command while in PvP combat");
     }
 
     protected void requireNonFlyingPlayer (EntityPlayer player)
@@ -85,7 +107,7 @@ public abstract class AbstractPvPCommand extends CommandBase
     protected void requireNonPvPPlayer (EntityPlayer player)
     {
         if (PvPUtils.isInPvP (player))
-            throw new CommandException ("You cannot use this command while that player is in PvP");
+            throw new CommandException ("You cannot use this command while that player is in PvP combat");
     }
 
     protected void requireSenderWithPvPEnabled (EntityPlayer sender)