+++ src/main/java/pvpmode/command/PvPCommand.java
@@ -1,5 +1,9 @@
 package pvpmode.command;
 
+import java.util.*;
+
+import org.apache.commons.lang3.tuple.Triple;
+
 import net.minecraft.command.ICommandSender;
 import net.minecraft.entity.player.*;
 import net.minecraft.event.ClickEvent;
@@ -19,6 +23,42 @@ public class PvPCommand extends AbstractPvPCommand
     public String getCommandUsage (ICommandSender sender)
     {
         return "/pvp [cancel] OR /pvp spy [on|off]";
+    }
+
+    @Override
+    public Collection<Triple<String, String, String>> getShortHelpMessages (ICommandSender sender)
+    {
+        Collection<Triple<String, String, String>> messages = new ArrayList<> ();
+        messages.add (Triple.of ("pvp", "", "Starts the warmup timer to toggle PvP."));
+        messages.add (Triple.of ("pvp cancel", "", "Cancels the warmup timer."));
+        if (PvPMode.allowPerPlayerSpying)
+        {
+            messages.add (Triple.of ("pvp spy ", "[on|off]", "Toggles the spy settings."));
+        }
+        return messages;
+    }
+
+    @Override
+    public Collection<Triple<String, String, String>> getLongHelpMessages (ICommandSender sender)
+    {
+        Collection<Triple<String, String, String>> messages = new ArrayList<> ();
+        messages.add (
+            Triple.of ("pvp", "",
+                "Starts your warmup timer. After the timer runs out, your PvP mode will be toggled. Your PvP mode mustn't be overridden."));
+        messages.add (Triple.of ("pvp cancel", "",
+            "Cancels your warmup timer, if it's running. Your PvP mode mustn't be overridden."));
+        if (PvPMode.allowPerPlayerSpying)
+        {
+            messages.add (Triple.of ("pvp spy ", "[on|off]",
+                "Allows you to either toggle your spy settings, or to set them to a specific value (on or off). If spying is enabled for you, you can receive proximity and direction information about other players with PvP enabled (via the PvP list). PvP needs to be enabled for you."));
+        }
+        return messages;
+    }
+
+    @Override
+    public String getGeneralHelpMessage (ICommandSender sender)
+    {
+        return "Allows you to manage your PvP mode (OFF, ON without spy, ON with spy). You mustn't be able to fly, not in creative mode and not in PvP combat to use this command.";
     }
 
     @Override