+++ CHANGELOG.md
@@ -22,6 +22,9 @@
 * Unified the style and content of the chat messages
 * Admins can now specify with `pvpadmin <player> [on|off]` the new PvP mode of a player
 * A message will now inform the player if the warmup timer was canceled because of PvP
+* Made some chat messages more clear
+* Added `pvphelp <commandName>` to retrieve detailed information about the specified command
+* Restructured the whole `pvphelp` overview menu
 * Global chat messages are now prefixed with a configurable prefix
 
 ### LOTR compatibility: